+++ app/src/main/java/shevtsov/daniil/incrementalreader/creation/presentation/CreationScreenEvent.kt
@@ -1,4 +1,4 @@
-package shevtsov.daniil.incrementalreader.creation
+package shevtsov.daniil.incrementalreader.creation.presentation
 
 sealed class CreationScreenEvent {
