--- app/src/main/java/shevtsov/daniil/incrementalreader/learning/LearningViewModel.kt
@@ -1,10 +1,0 @@
-package shevtsov.daniil.incrementalreader.learning
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class LearningViewModel @Inject constructor(
-
-) : ViewModel() {
-
-}
