+++ app/src/main/java/shevtsov/daniil/incrementalreader/learning/presentation/LearningScreenEvent.kt
@@ -1,0 +1,4 @@
+package shevtsov.daniil.incrementalreader.learning.presentation
+
+class LearningScreenEvent {
+}
