+++ app/src/main/java/shevtsov/daniil/incrementalreader/learning/presentation/LearningViewState.kt
@@ -1,0 +1,6 @@
+package shevtsov.daniil.incrementalreader.learning.presentation
+
+data class LearningViewState(
+    val itemName: String,
+    val itemContent: String,
+)
