+++ app/src/main/java/shevtsov/daniil/incrementalreader/main/presentation/MainScreenEvent.kt
@@ -1,4 +1,4 @@
-package shevtsov.daniil.incrementalreader.main
+package shevtsov.daniil.incrementalreader.main.presentation
 
 sealed class MainScreenEvent {
