+++ app/src/main/java/shevtsov/daniil/incrementalreader/main/presentation/MainViewModel.kt
@@ -1,4 +1,4 @@
-package shevtsov.daniil.incrementalreader.main
+package shevtsov.daniil.incrementalreader.main.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
