+++ app/src/main/java/shevtsov/daniil/incrementalreader/storage/domain/GetSavedItemUseCase.kt
@@ -1,0 +1,11 @@
+package shevtsov.daniil.incrementalreader.storage.domain
+
+import javax.inject.Inject
+
+class GetSavedItemUseCase @Inject constructor(
+    private val repository: TextStorageRepository
+) {
+
+    operator fun invoke(itemId: String) = repository.getItem(itemId = itemId)
+
+}
