+++ app/src/main/java/shevtsov/daniil/incrementalreader/storage/domain/TextStorageRepository.kt
@@ -6,4 +6,6 @@ interface TextStorageRepository {
 
     fun getItems(): List<InformationItem>
 
+    fun getItem(itemId: String): InformationItem?
+
 }
