--- app/src/main/java/shevtsov/daniil/incrementalreader/structure/StructureViewState.kt
@@ -1,5 +1,0 @@
-package shevtsov.daniil.incrementalreader.structure
-
-data class StructureViewState(
-    val items: List<StructureInformationItem>
-)
