+++ app/src/main/java/shevtsov/daniil/incrementalreader/structure/presentation/StructureInformationItemMapper.kt
@@ -1,6 +1,7 @@
-package shevtsov.daniil.incrementalreader.structure
+package shevtsov.daniil.incrementalreader.structure.presentation
 
 import shevtsov.daniil.incrementalreader.storage.domain.InformationItem
+import shevtsov.daniil.incrementalreader.structure.view.adapter.StructureInformationItem
 import javax.inject.Inject
 
 class StructureInformationItemMapper @Inject constructor() {
