+++ app/src/main/java/shevtsov/daniil/incrementalreader/structure/presentation/StructureScreenEvent.kt
@@ -1,4 +1,4 @@
-package shevtsov.daniil.incrementalreader.structure
+package shevtsov.daniil.incrementalreader.structure.presentation
 
 sealed class StructureScreenEvent {
     data class OpenLearning(val itemId: String) : StructureScreenEvent()
