+++ app/src/main/java/shevtsov/daniil/incrementalreader/structure/presentation/StructureViewState.kt
@@ -1,0 +1,7 @@
+package shevtsov.daniil.incrementalreader.structure.presentation
+
+import shevtsov.daniil.incrementalreader.structure.view.adapter.StructureInformationItem
+
+data class StructureViewState(
+    val items: List<StructureInformationItem>
+)
