+++ app/src/main/java/shevtsov/daniil/incrementalreader/structure/view/adapter/StructureAdapterAction.kt
@@ -1,4 +1,4 @@
-package shevtsov.daniil.incrementalreader.structure.adapter
+package shevtsov.daniil.incrementalreader.structure.view.adapter
 
 sealed class StructureAdapterAction {
     data class ItemSelected(val itemId: String) : StructureAdapterAction()
